refactor(padi): split the UI-chrome vocabulary into chromeVocab.ts (L17)
vocab.ts braided two volatility axes in one ~960-line module: the PTY-lifecycle /
session-persistence schemas and the UI-chrome schemas (canvas layout, sub-panel,
Code-tab views, right-panel state) plus their presentation helpers (viewLabel,
CODE_TAB_VIEW_ORDER, rightPanelView) — a canvas tweak and a PTY change churned the
same file, and presentation logic lived in the vocabulary. Home the chrome half in
a sibling chromeVocab.ts. The terminal-metadata schemas consume the three chrome
schemas, so the arrow points one way (vocab.ts -> chromeVocab.ts, never back).

Same export surface: surface.ts re-exports both modules, so @kolu/padi/surface
carries the identical set — no consumer import changes. chromeVocab.ts is zod-only
and browser-safe; the package seal's reverse arms stay green.
